fix: prefer makeStyle(theme => ...) over using theme directly
One less import, so less code.
