+++ app/frontend/src/features/profile/constants.ts
@@ -1,0 +1,10 @@
+import { SmokingLocation } from "../../pb/api_pb";
+
+export const smokingLocationLabels = {
+  [SmokingLocation.SMOKING_LOCATION_NO]: "No",
+  [SmokingLocation.SMOKING_LOCATION_OUTSIDE]: "Outside",
+  [SmokingLocation.SMOKING_LOCATION_WINDOW]: "Window",
+  [SmokingLocation.SMOKING_LOCATION_YES]: "Yes",
+  [SmokingLocation.SMOKING_LOCATION_UNSPECIFIED]: "",
+  [SmokingLocation.SMOKING_LOCATION_UNKNOWN]: "Unknown",
+};
